+++ examples/example.py
@@ -1,20 +1,19 @@
 from pyautocv.segmentation import *
 import os
-
 os.chdir("pyautocv")
 # gray images
-images_list = Segmentation("images/cats")
-show_images(gray_images(images_list.read_images()), images_list.read_images())
+images_list = Segmentation("images/dic", image_suffix="tif", color_mode="gray")
+show_images(images_list.read_images(), images_list.read_images(), number=10)
 # smooth images
 show_images(images_list.read_images(), images_list.smooth())
 # use median
-show_images(images_list.read_images(), images_list.smooth(mask="median", kernel_shape=(7, 7)))
+show_images(images_list.read_images(), images_list.smooth(mask="median", kernel_shape=(7, 7)), number = 8)
 
 # detect edges
 
-show_images(images_list.read_images(), images_list.detect_edges(operator="roberts", mask="gaussian", sigma=0.8))
+show_images(images_list.read_images(), images_list.detect_edges(operator="roberts", mask="gaussian", sigma=0.8), number=4)
 
-show_images(images_list.read_images(), images_list.detect_edges(operator="laplace", mask="gaussian", sigma=0.8))
+show_images(images_list.read_images(), images_list.threshold_images(), number = 2)
 
 # threshold
 cats = Segmentation("images/cats")